requirements.txt
tencentcloud-sdk-python
python-dotenv
pytz
requests
mysql-connector-python 